Docs: don't commit the generated enums.md to the repo
Just generate it as we build the docs
