+++ .github/workflows/build_docs.yaml
@@ -69,15 +69,6 @@ jobs:
         tool_name: formatters
         level: error
         fail_on_error: true
-    - name: Check Enum docs
-      run: cargo xtask enumdocs
-    - name: Suggest enum doc changes
-      uses: reviewdog/action-suggester@v1
-      with:
-        tool_name: xtask enumdocs
-        level: error
-        fail_on_error: true
-
     - name: Check license headers
       run: cargo xtask check_license_headers --fix-it || echo "License information needs an update"
     - name: Suggest license header changes
