+++ xtask/src/slintdocs.rs
@@ -5,6 +5,7 @@
 
 use anyhow::{Context, Result};
 use std::ffi::OsString;
+use std::io::Write;
 use std::path::{Path, PathBuf};
 
 fn symlink_file<P: AsRef<Path>, Q: AsRef<Path>>(src: P, dst: Q) -> Result<()> {
@@ -56,6 +57,8 @@ fn symlink_files_in_dir<S: AsRef<Path>, T: AsRef<Path>, TS: AsRef<Path>>(
 }
 
 pub fn generate(show_warnings: bool) -> Result<(), Box<dyn std::error::Error>> {
+    generate_enum_docs()?;
+
     let root = super::root_dir();
 
     let docs_source_dir = root.join("docs/language");
@@ -98,6 +101,74 @@ pub fn generate(show_warnings: bool) -> Result<(), Box<dyn std::error::Error>> {
 
     if show_warnings {
         println!("{}", String::from_utf8_lossy(&output.stderr));
+    }
+
+    Ok(())
+}
+
+pub fn generate_enum_docs() -> Result<(), Box<dyn std::error::Error>> {
+    let mut enums: std::collections::BTreeMap<String, String> = std::collections::BTreeMap::new();
+
+    macro_rules! gen_enums {
+        ($( $(#[doc = $enum_doc:literal])* $(#[non_exhaustive])? enum $Name:ident { $( $(#[doc = $value_doc:literal])* $Value:ident,)* })*) => {
+            $(
+                let mut entry = format!("## `{}`\n\n", stringify!($Name));
+                $(entry += &format!("{}\n", $enum_doc);)*
+                entry += "\n";
+                $(
+                    let mut has_val = false;
+                    entry += &format!("* **`{}`**:", to_kebab_case(stringify!($Value)));
+                    $(
+                        if has_val {
+                            entry += "\n   ";
+                        }
+                        entry += &format!("{}", $value_doc);
+                        has_val = true;
+                    )*
+                    entry += "\n";
+                )*
+                entry += "\n";
+                enums.insert(stringify!($Name).to_string(), entry);
+            )*
+        }
+    }
+
+    #[allow(unused)] // for 'has_val'
+    {
+        i_slint_common::for_each_enums!(gen_enums);
+    }
+
+    let root = super::root_dir();
+
+    let path = root.join("docs/language/src/builtins/enums.md");
+    let mut file = std::fs::File::create(&path).context(format!("error creating {path:?}"))?;
+
+    file.write_all(
+        br#"<!-- Generated with `cargo xtask slintdocs` from internal/commons/enums.rs -->
+# Builtin Enumerations
+
+"#,
+    )?;
+
+    for (_, v) in enums {
+        // BTreeMap<i64, String>
+        write!(file, "{v}")?;
+    }
+
+    /// Convert a ascii pascal case string to kebab case
+    fn to_kebab_case(str: &str) -> String {
+        let mut result = Vec::with_capacity(str.len());
+        for x in str.as_bytes() {
+            if x.is_ascii_uppercase() {
+                if !result.is_empty() {
+                    result.push(b'-');
+                }
+                result.push(x.to_ascii_lowercase());
+            } else {
+                result.push(*x);
+            }
+        }
+        String::from_utf8(result).unwrap()
     }
 
     Ok(())
